test_dspy/pyproject.toml
[tool.poetry]
name = "test-dspy"
version = "0.1.0"
description = ""
authors = ["sparktsao <linhots.tsao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dspy-ai = "^2.6.27"
dotenv = "^0.9.9"
python-dotenv = "^1.1.1"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
